// src/generated/display.ts
/**
 * This file was automatically generated by joi-to-typescript
 * Do not modify this file manually
 */

export interface Display {
  convertedImage?: any;
  imageSettings: ImageSettings;
  isGeneratedFromDefaultBackImage: boolean;
  originalImage?: any;
  previewImage: string;
  textSettings: TextSettings;
  textWithIconSettings: TextWithIconSettings;
}

export interface ImageSettings {
  autoCrop: boolean;
  blackThreshold: number;
  brightness: number;
  contrast: number;
  imageMode: 'dither' | 'normal' | 'hybrid';
  invert: boolean;
  whiteThreshold: number;
}

export interface TextSettings {
  font: 'fonts/smaller.fnt' | 'fonts/small.fnt' | 'fonts/medium.fnt' | 'fonts/large.fnt';
  position: 'bottom' | 'right';
  text?: string;
}

export interface TextWithIconSettings {
  iconWidthMultiplier: number;
}